requirements.txt
groq==0.37.0
langchain==1.1.0
langchain_core==1.1.0
langchain_google_genai==3.2.0
librosa==0.11.0
nltk==3.9.1
noisereduce==3.0.3
numpy==2.3.5
openai_whisper==20250625
pandas==2.3.3
protobuf==6.33.1
pydantic==2.12.5
python-dotenv==1.2.1
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
sounddevice==0.5.3
sounddevice==0.5.2
soundfile==0.13.1
streamlit==1.50.0
streamlit==1.42.0
